adding bms status
